backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
transformers==4.38.2
Pillow==10.2.0
requests==2.31.0
faiss-cpu==1.7.4
torch==2.1.2
open-clip-torch==2.24.0
numpy==1.26.4
jupyter==1.1.1